ClaudeBackend/sistema-precios-lafuga/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-multipart==0.0.6
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
httpx==0.26.0
